Stage 1: Server thread, Config and forwarding structures
At this stage, the program can display help and can start a server thread
that does nothing at the moment.

The forwarder config file is parsed and put into the optimized data structure.

The configured HMC is contacted and the CPCs and LPARs/Partitions from the
config file are listed.

For demo purposes, the program prints whether each existing partition matches
the forwarding config.

Signed-off-by: Andreas Maier <maiera@de.ibm.com>