Stage 2: Subscribe for OS messages and display them
In stage 2, the OS message channels are opened for the LPARs/partitions
to be forwarded, and the incoming notifications are displayed.

For later use, the object notifications are also subscribd for, but
that will need to be changed to become inventory notifications
instead, in order to track partitions that come and go.

Signed-off-by: Andreas Maier <maiera@de.ibm.com>